renamed authHelper.ts to auth-helper.ts
