+++ apps/frontend/lib/auth-helpers.ts
